backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.1
chromadb==0.4.18
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
wikipedia==1.4.0
newsapi-python==0.2.7
praw==7.7.1
langsmith==0.0.78
rouge-score==0.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
langchain-core
